# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeVault API",
        "description": "OpenAPI specification for the AnimeVault backend",
        "contact": {
            "email": "support@animevault.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/me/notifications": {
            "get": {
                "tags": [
                    "User Notifications"
                ],
                "summary": "Lists the authenticated user notifications",
                "operationId": "apiMeNotificationsIndex",
                "responses": {
                    "200": {
                        "description": "User notifications list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/notifications/{notification}/read": {
            "post": {
                "tags": [
                    "User Notifications"
                ],
                "summary": "Marks a single authenticated user notification as read",
                "operationId": "apiMeNotificationsRead",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "description": "Notification identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/notifications/read-all": {
            "post": {
                "tags": [
                    "User Notifications"
                ],
                "summary": "Marks all authenticated user notifications as read",
                "operationId": "apiMeNotificationsReadAll",
                "responses": {
                    "200": {
                        "description": "Notifications marked as read"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctumBearerAuth": {
                "type": "http",
                "description": "Bearer token returned by the Sanctum login or register endpoints",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}